--- frontend/src/pages/host/AuthCallBack/styles.ts
@@ -1,13 +1,0 @@
-import { css } from '@emotion/react';
-
-const layout = css`
-  display: flex;
-  height: 100vh;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-const styles = { layout };
-
-export default styles;
